int => decimal(10,2) in tables

+++ CoCowork.DB/Tables/MiniOffice.sql
@@ -3,7 +3,7 @@
 	[Id] INT NOT NULL PRIMARY KEY IDENTITY, 
     [Name] VARCHAR(30) NOT NULL, 
     [AmountOfPlaces] INT NOT NULL, 
-    [PricePerDay] INT NOT NULL, 
+    [PricePerDay] DECIMAL(10, 2) NOT NULL, 
     [IsActive] BIT NOT NULL,
 
 )
